ai-music-comparison/requirements-bark.txt
accelerate==1.11.0
boto3==1.40.69
botocore==1.40.69
certifi==2025.10.5
charset-normalizer==3.4.4
einops==0.8.1
encodec==0.1.1
filelock==3.20.0
fsspec==2025.10.0
funcy==2.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
packaging==25.0
pillow==12.0.0
psutil==7.1.3
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
s3transfer==0.14.0
safetensors==0.6.2
scipy==1.16.3
six==1.17.0
suno-bark @ git+https://github.com/suno-ai/bark.git@f4f32d4cd480dfec1c245d258174bc9bde3c2148
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1
typing_extensions==4.15.0
urllib3==2.5.0
